Units are spawned and they move towards the enemy tower
Units are randomly spawned in a given spawn plane area. These units determine which of the 3 bridges is closest to them, then they cross that bridge and move towards the enemy tower.
All the important changes are made to WayPointFollower and Spawner scripts only. The scripts can be found under Assets/Scripts.